trycp: Extract module `app_interface.rs`
