+++ crates/trycp_server/Cargo.toml
@@ -23,9 +23,8 @@ once_cell = "1.5.0"
 parking_lot = "0.10.0"
 reqwest = "=0.9.11"
 rmp-serde = "0.14.3"
-serde = "=1.0.104"
+serde = { version = "=1.0.104", features = ["derive"] }
 serde_bytes = "0.11.0"
-serde_derive = "=1.0.104"
 serde_json = { version = "=1.0.47", features = ["preserve_order"] }
 slab = "0.4.0"
 structopt = "=0.2.18"
